Kolejne poprawki
