+++ programyPython/algorytmMapowania/Package/__init__.py
@@ -1,2 +1,3 @@
 from mapMatrixClass import mapMatrixClass
+from simulationMapMatrixClass import simulationMapMatrixClass
 from robotClass import robotClass
