+++ programyPython/algorytmMapowania/Package/mapMatrixClass.py
@@ -1,6 +1,5 @@
 import numpy as np
 import sys
-import random
 
 class mapMatrixClass():
 	def __init__(self,x,y):
@@ -8,187 +7,21 @@ class mapMatrixClass():
 		self.ySize = y
 		self.mapMatrix = np.zeros((self.xSize,self.ySize),dtype=np.byte)
 
-	def prepareTable(self):		#metoda tylko do testow tylko dla tablicy 10X10
-		self.setSolidWall(0,0,"W")
-		self.setSolidWall(0,0,"S")
-		self.setSolidWall(0,1,"W")
-		self.setSolidWall(0,2,"W")
-		self.setSolidWall(0,3,"W")
-		self.setSolidWall(0,4,"W")
-		self.setSolidWall(0,5,"W")
-		self.setSolidWall(0,6,"W")
-		self.setSolidWall(0,6,"N")
-		self.setSolidWall(0,7,"W")
-		self.setSolidWall(0,7,"S")
-		self.setSolidWall(0,8,"W")
-		self.setSolidWall(0,9,"W")
-		self.setSolidWall(0,9,"N")
-
-		self.setSolidWall(1,0,"S")
-		self.setSolidWall(1,6,"N")
-		self.setSolidWall(1,7,"S")
-		self.setSolidWall(1,7,"E")
-		self.setSolidWall(1,8,"E")
-		self.setSolidWall(1,9,"N")
-
-		self.setSolidWall(2,0,"S")
-		self.setSolidWall(2,5,"E")
-		self.setSolidWall(2,6,"E")
-		self.setSolidWall(2,7,"W")
-		self.setSolidWall(2,8,"W")
-		self.setSolidWall(2,8,"N")
-		self.setSolidWall(2,9,"S")
-		self.setSolidWall(2,9,"N")
-
-		self.setSolidWall(3,0,"S")
-		self.setSolidWall(3,2,"E")
-		self.setSolidWall(3,4,"N")
-		self.setSolidWall(3,5,"S")
-		self.setSolidWall(3,5,"W")
-		self.setSolidWall(3,6,"W")
-		self.setSolidWall(3,6,"N")
-		self.setSolidWall(3,6,"E")
-		self.setSolidWall(3,7,"S")
-		self.setSolidWall(3,8,"N")
-		self.setSolidWall(3,8,"E")
-		self.setSolidWall(3,9,"N")
-		self.setSolidWall(3,9,"S")
-
-		self.setSolidWall(4,0,"S")
-		self.setSolidWall(4,1,"E")
-		self.setSolidWall(4,2,"N")
-		self.setSolidWall(4,2,"W")
-		self.setSolidWall(4,3,"S")
-		self.setSolidWall(4,4,"N")
-		self.setSolidWall(4,5,"S")
-		self.setSolidWall(4,6,"W")
-		self.setSolidWall(4,6,"E")
-		self.setSolidWall(4,7,"N")
-		self.setSolidWall(4,8,"E")
-		self.setSolidWall(4,8,"S")
-		self.setSolidWall(4,8,"W")
-		self.setSolidWall(4,9,"E")
-		self.setSolidWall(4,9,"N")
-
-		self.setSolidWall(5,0,"S")
-		self.setSolidWall(5,0,"N")
-		self.setSolidWall(5,1,"S")
-		self.setSolidWall(5,1,"E")
-		self.setSolidWall(5,1,"W")
-		self.setSolidWall(5,2,"N")
-		self.setSolidWall(5,2,"E")
-		self.setSolidWall(5,3,"S")
-		self.setSolidWall(5,4,"N")
-		self.setSolidWall(5,5,"S")
-		self.setSolidWall(5,5,"E")
-		self.setSolidWall(5,5,"N")
-		self.setSolidWall(5,6,"N")
-		self.setSolidWall(5,6,"E")
-		self.setSolidWall(5,6,"S")
-		self.setSolidWall(5,6,"W")
-		self.setSolidWall(5,7,"S")
-		self.setSolidWall(5,8,"W")
-		self.setSolidWall(5,9,"N")
-		self.setSolidWall(5,9,"W")
-
-		self.setSolidWall(6,0,"S")
-		self.setSolidWall(6,1,"E")
-		self.setSolidWall(6,1,"W")
-		self.setSolidWall(6,2,"E")
-		self.setSolidWall(6,2,"W")
-		self.setSolidWall(6,5,"W")
-		self.setSolidWall(6,6,"W")
-		self.setSolidWall(6,9,"N")
-
-		self.setSolidWall(7,0,"S")
-		self.setSolidWall(7,0,"E")
-		self.setSolidWall(7,1,"E")
-		self.setSolidWall(7,1,"W")
-		self.setSolidWall(7,2,"W")
-		self.setSolidWall(7,2,"N")
-		self.setSolidWall(7,3,"N")
-		self.setSolidWall(7,3,"S")
-		self.setSolidWall(7,3,"E")
-		self.setSolidWall(7,4,"S")
-		self.setSolidWall(7,7,"E")
-		self.setSolidWall(7,9,"N")
-
-		self.setSolidWall(8,0,"S")
-		self.setSolidWall(8,0,"W")
-		self.setSolidWall(8,1,"W")
-		self.setSolidWall(8,1,"N")
-		self.setSolidWall(8,2,"S")
-		self.setSolidWall(8,2,"E")
-		self.setSolidWall(8,3,"E")
-		self.setSolidWall(8,3,"W")
-		self.setSolidWall(8,4,"E")
-		self.setSolidWall(8,6,"E")
-		self.setSolidWall(8,6,"N")
-		self.setSolidWall(8,7,"N")
-		self.setSolidWall(8,7,"S")
-		self.setSolidWall(8,7,"W")
-		self.setSolidWall(8,8,"S")
-		self.setSolidWall(8,9,"N")
-
-		self.setSolidWall(9,0,"S")
-		self.setSolidWall(9,0,"E")
-		self.setSolidWall(9,1,"E")
-		self.setSolidWall(9,2,"E")
-		self.setSolidWall(9,2,"W")
-		self.setSolidWall(9,3,"W")
-		self.setSolidWall(9,3,"E")
-		self.setSolidWall(9,4,"E")
-		#self.setSolidWall(9,4,"N")
-		self.setSolidWall(9,4,"W")
-		self.setSolidWall(9,5,"E")
-		#self.setSolidWall(9,5,"S")
-		self.setSolidWall(9,5,"N")
-		self.setSolidWall(9,6,"S")
-		self.setSolidWall(9,6,"E")
-		self.setSolidWall(9,6,"W")
-		self.setSolidWall(9,7,"N")
-		self.setSolidWall(9,7,"E")
-		self.setSolidWall(9,8,"S")
-		self.setSolidWall(9,8,"E")
-		self.setSolidWall(9,9,"E")
-		self.setSolidWall(9,9,"N")
-
-	def generateRandomMap(self):
-		self.generateMapEdge()
-		for i in range(0,self.xSize):
-			for j in range(0,self.ySize):
-				if(random.randint(0,1) == 1):
-					self.setSolidWall(i,j,"N")
-					if(i != self.xSize-1):
-						self.setSolidWall(i+1,j,"S")
-				if(random.randint(0,1) == 1):
-					 self.setSolidWall(i,j,"E")
-					 if(j != self.ySize-1):
-						 self.setSolidWall(i,j+1,"W")
-				if(random.randint(0,1) == 1):
-					 self.setSolidWall(i,j,"S")
-					 if(i != 0):
-						 self.setSolidWall(i-1,j,"N")
-				if(random.randint(0,1) == 1):
-					self.setSolidWall(i,j,"W")
-					if(j !=0):
-						self.setSolidWall(i,j-1,"E")
-
-	def generateMapEdge(self):
-		for i in range(0,self.xSize):
-			self.setSolidWall(i,self.ySize-1,"N")
-			self.setSolidWall(i,0,"S")
-		for i in range(0,self.ySize):
-			self.setSolidWall(0,i,"W")
-			self.setSolidWall(self.xSize-1,i,"E")
-
 	def setSolidWall(self,x,y,direction):
-		if direction == "N" : self.mapMatrix[x][y] = self.mapMatrix[x][y] | 8
-		elif direction == "E" : self.mapMatrix[x][y] = self.mapMatrix[x][y] | 4
-		elif direction == "S" : self.mapMatrix[x][y] = self.mapMatrix[x][y] | 2
-		elif direction == "W" : self.mapMatrix[x][y] = self.mapMatrix[x][y] | 1
+		self.checkVariablesForMatrix(x,y)
+		if direction == "N":
+			self.mapMatrix[x][y] = self.mapMatrix[x][y] | 8
+		elif direction == "E":
+			 self.mapMatrix[x][y] = self.mapMatrix[x][y] | 4
+		elif direction == "S":
+			self.mapMatrix[x][y] = self.mapMatrix[x][y] | 2
+		elif direction == "W":
+			self.mapMatrix[x][y] = self.mapMatrix[x][y] | 1
+		else:
+			raise Exception("Wrong parameter for direction")
 
 	def isWallExist(self,x,y,direction):
+		self.checkVariablesForMatrix(x,y)
 		if direction == "N" :
 			if self.mapMatrix[x][y] & 8 == 8: return True
 		elif direction == "E" :
@@ -197,21 +30,31 @@ class mapMatrixClass():
 			if self.mapMatrix[x][y] & 2 == 2: return True
 		elif direction == "W" :
 			if self.mapMatrix[x][y] & 1 == 1: return True
+		else:
+			raise Exception("Wrong parameter for direction")
 		return False
 
 	def setVisited(self,x,y):
+		self.checkVariablesForMatrix(x,y)
 		self.mapMatrix[x][y] = self.mapMatrix[x][y] | 16
 
 	def wasVisited(self,x,y):
+		self.checkVariablesForMatrix(x,y)
 		if self.mapMatrix[x][y] & 16 == 16: return True
 		else: return False
 
 	def setNotAvailable(self,x,y):
+		self.checkVariablesForMatrix(x,y)
 		self.mapMatrix[x][y] = self.mapMatrix[x][y] | 32
 
 	def isNotAvailable(self,x,y):
+		self.checkVariablesForMatrix(x,y)
 		if self.mapMatrix[x][y] & 32 == 32: return True
 		else: return False
+
+	def checkVariablesForMatrix(self,x,y):
+		if(x < 0 or x >= self.xSize or y < 0 or y >= self.ySize):
+			raise Exception("Matrix indexes outside the scope")
 
 	def findWayToNearestNoVisitedSpot(self,robotXposition,robotYposition):
 		pathMap = np.zeros((self.xSize,self.ySize),dtype=np.byte)
